+++ src/Symfony/Bridge/Validator/Constraints/IdCard.php
@@ -15,10 +15,6 @@ namespace Ajgl\ValidatorEs\Symfony\Bridge\Validator\Constraints;
 
 use Symfony\Component\Validator\Constraint;
 
-/**
- * @Annotation
- * @Target({"PROPERTY", "METHOD", "ANNOTATION"})
- */
 #[\Attribute(\Attribute::TARGET_PROPERTY | \Attribute::TARGET_METHOD | \Attribute::IS_REPEATABLE)]
 final class IdCard extends Constraint
 {