+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Symfony 5.4 support
+
 ## [0.1.8] - 2026-06-01
 
 ### Removed